CHROMIUM: ecryptfs: Have the kthread open the lower file under SELinux
When SELinux is enabled, files are associated with the current task's
SID when they are opened. Before this change, the lower file would
sometimes be associated with the SID of the first process to open the
corresponding eCryptfs file. If a second process attempts to access this
file through the same lower file, the access is denied unless the
SELinux policy allows the second process to use files opened by the
first process (the FD__USE permission).

Since it may not be desirable to allow all processes this permission,
this change always opens the lower file from a thread with the kernel
SID. This way, the policy only needs to grant all processes FD__USE on
the kernel context (which in practice, is exactly the use of eCryptfs
lower files).

Signed-off-by: Ricky Zhou <rickyz@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/337349
Reviewed-by: Gwendal Grignou <gwendal@google.com>
(cherry picked from commit 822ea6dec86d9abdfbb3123aa1b5d5536ee77f12)
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>

BUG=b:25408463
BUG=chrome-os-partner:55277
TEST=Play store works

Change-Id: If22e125a48f474b7f0dfeb162bc77bd603431ba1
Reviewed-on: https://chromium-review.googlesource.com/362714
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Ricky Zhou <rickyz@chromium.org>