FROMLIST: platform: chrome: cros_ec_dev - Return number of bytes read with CROS_EC_DEV_IOCRDMEM
While ioctls normally return a negative error or 0 on success, this is not
the case for CROS_EC_DEV_IOCXCMD in the cros_ec driver, which returns the
number of bytes read instead. However, CROS_EC_DEV_IOCRDMEM in the same
driver does not return the number of byte read on success. This is both
inconsistent and problematic, since the amount of returned data can be
dynamic. Have it return the number of data bytes for consistency.

BUG=b:73230881
TEST=Run "ectool battery" on affected system

Change-Id: I4d839a3e0b73ede537ec81cd52ca2a0dd9d21c44
Cc: Gwendal Grignou <gwendal@chromium.org>
Signed-off-by: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(am from https://patchwork.kernel.org/patch/8754601/)
Reviewed-on: https://chromium-review.googlesource.com/914844
Reviewed-by: Benson Leung <bleung@chromium.org>