CHROMIUM: cros_ec: add a poll handler
Allow to poll on the cros_ec device to receive the MKBP events.
The /dev/cros_{ec|fp|..} file operations now implements the poll
operation.
The userspace can now receive specific MKBP events by doing the
following:
- open the /dev/cros_XX file.
- call the CROS_EC_DEV_IOCEVENTMASK ioctl with the bitmap of the MKBP
  events it wishes to receive as argument.
- poll on the file descriptor.
- when it gets POLLIN, do a read on the file descriptor, the first
  queued event will be returned (using the struct
  ec_response_get_next_event format: one byte of event type, then
  the payload).

The read() operation returns at most one event even if there are several
queued, and it might be truncated if the buffer is smaller than the
event (but the caller should know the maximum size of the events it is
reading).
read() used to return the EC version string, it still does it when no
event mask or an empty event is set for backward compatibility (despite
nobody really using this feature).

This will be used by the userspace daemon to receive and treat the
EC_MKBP_EVENT_FINGERPRINT  sent by the FP MCU.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=b:69460856, b:71986991
TEST=manual, run on Eve EVT and generate FP MCU events with an updated
ectool.

Change-Id: I671d530b4139651275a3c56770c38fa8fc697543
Reviewed-on: https://chromium-review.googlesource.com/806115
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>

Conflicts:
	drivers/mfd/cros_ec_dev.c

[rebase419(groeck): Context conflicts]
Signed-off-by: Guenter Roeck <groeck@chromium.org>