PM / Domain: Implement of_dev_pm_genpd_opp_to_performance_state()
This implements of_dev_pm_genpd_opp_to_performance_state() which can be
used from the device drivers or the OPP core to find the performance
state encoded in the "required-opps" property of a node. Normally this
would be called only once for each OPP of the device for which the OPP
table of the device is getting generated.

Different platforms may encode the performance state differently using
the OPP table (they may simply return value of opp-hz or opp-microvolt,
or apply some algorithm on top of those values) and so a new callback
->opp_to_performance_state() is implemented to allow platform specific
drivers to convert the power domain OPP to a performance state value.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Acked-by: Ulf Hansson <ulf.hansson@linaro.org>