x86 / ACPI / cstate: Optimize C3 entry on Centaur CPUs
For Centaur CPUs the ucode will take care of the preservation of cache
coherence between CPU cores in C-states and the other CPU cores
regardless of how deep the C-states are, so it is not necessary to
flash the caches in software before entering C3.

Also the bus master arbitration need not be disabled on C3 entry for
these CPUs, because the chipset will take care of it automatically on
all Centaur platforms.

Signed-off-by: David Wang <davidwang@zhaoxin.com>
[ rjw: Subject/changelog, comments ]
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>