mm: memcg: plumbing memcg for kmalloc allocations
Introduce the memcg variant for kmalloc allocation functions.  The kmalloc
allocations are underlying served using the kmem caches unless the size of
the allocation request is larger than KMALLOC_MAX_CACHE_SIZE, in which
case, the kmem caches are bypassed and the request is routed directly to
page allocator.  So, for __GFP_ACCOUNT kmalloc allocations, the memcg of
current task is charged.  This patch introduces memcg variant of kmalloc
functions to allow callers to provide memcg for charging.

Link: http://lkml.kernel.org/r/20180221030101.221206-3-shakeelb@google.com
Signed-off-by: Shakeel Butt <shakeelb@google.com>
Cc: Amir Goldstein <amir73il@gmail.com>
Cc: Christoph Lameter <cl@linux.com>
Cc: David Rientjes <rientjes@google.com>
Cc: Greg Thelen <gthelen@google.com>
Cc: Jan Kara <jack@suse.cz>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Mel Gorman <mgorman@suse.de>
Cc: Michal Hocko <mhocko@kernel.org>
Cc: Pekka Enberg <penberg@kernel.org>
Cc: Vladimir Davydov <vdavydov.dev@gmail.com>
Cc: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>