lge_battery: Use -EPROBE_DEFER instead of -ENODEV during probe
When this driver is built into the kernel image, bm_init fails because
the parallel power supply is not up yet, which causes the probe routine
to fail.

[    1.375729] lge_battery: bm_init: Couldn't get pl_psy
[    1.375736] lge_battery: lge_battery_probe: bm_init fail

The kernel handles this by allowing the probe routine to be deferred
if the driver requests it by returning -EPROBE_DEFER, per a comment
in drivers/base/dd.c:

"If a required resource is not available yet, a driver can request
probing to be deferred by returning -EPROBE_DEFER from its probe hook"

Ensure the driver performs the same as both a built-in driver and
loadable kernel module.

Change-Id: Ia2e380da63c504b1ee0d7489c69a3acf0601f282
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>