mm: memfd: split out memfd for use by multiple filesystems
When memfd_create support was originally written, it only provided support
for tmpfs.  Hence, the code was added to files providing tmpfs
functionality and build when CONFIG_TMPFS was enabled.

memfd support has recently been added for hugetlbfs.  In an effort to make
it depend on tmpfs -or- hugetlbfs, split out the required memfd code to
separate files.

While moving code, fixed sparse warnings that existed in the original code
to keep kbuild test robot happy.

These files are not used until a subsequent patch which deletes duplicate
code in the orifinal files and enables their use.

Link: http://lkml.kernel.org/r/20180131022911.23947-3-mike.kravetz@oracle.com
Signed-off-by: Mike Kravetz <mike.kravetz@oracle.com>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: David Herrmann <dh.herrmann@gmail.com>
Cc: Hugh Dickins <hughd@google.com>
Cc: Khalid Aziz <khalid.aziz@oracle.com>
Cc: Marc-André Lureau <marcandre.lureau@gmail.com>
Cc: Michal Hocko <mhocko@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>